Fix a problem where old entries overwrite new ones
If a 1:2 split is change into a 2:1 split,
the old leaf nodes from the ":2" half of the old tree
are not cleared.
These are broadcasted to other processes.
If the 2:1 split is changed, again into a 1:2 split, these stale entries
become relevant because the nodes that don't repartition the subtrees
will clear the 2:1 subtree and, hence, not the ":2" part of the new
subtree.
If the old entries are larger than the new ones, the state of the tree
is inconsistent.

This commit fixes this problem by requiring the subtree repartition node
to clear the subtrees rooted at the children of the repartition root
first.